tweaking graphs for better responsiveness
In this commit, I have tried to align my graphs centrally, which fits the content on the site better. I have also removed a graph in which I wasn't using the correct data input to output the graph. I will have to log this in my 'left to implement' section on my readme file.